refactor: move current components following the new directory structure
